Factor out cluster actions and related state into ClusterActionsDropdown
